Revert changes made to ArmaContext
I'll make a new PR that restructures ArmaContext usage
