+++ arma-rs-example/src/lib.rs
@@ -1,6 +1,6 @@
 use arma_rs::{arma, Extension, Group};
 
-mod arma_info;
+mod arma_context;
 mod counter;
 mod system_info;
 mod timer;
@@ -22,7 +22,7 @@ fn init() -> Extension {
                 .command("french", welcome::french)
                 .command("spanish", welcome::spanish),
         )
-        .group("info", arma_info::group())
+        .group("context", arma_context::group())
         .group("counter", counter::group())
         .group("system", system_info::group())
         .group("timer", timer::group())
